add useAutoAdvanceState hook
